Updated the Device model to track favorite devices
Added a filled/empty star icon to use for the images on the PrimaryView
to show devices that have been favorited, and added a Toolbar button in
the Peripheral view to make allow the device to be favorited or
unfavorited. Other additions coming: favorite the device from
PrimaryView, favorited devices will always anchor at the top of the
device list if found, and scanning will happen automatically. The button
instead will be used to enable/disable connecting to all discovered
peripherals. And upon exiting the app, all connections will be canceled.