First attempt at using Combine.
Not sure it made a difference with regards to readability, but it was at
least useful in helping me separate out logic to manage device life cycle
away from the CentralManager code (if that was desirable).

Also removed a ton of code that I think I left in from my previous app that
was playing with Bluetooth.